Hero Section
Here I've made the Hero section of the website. I used the same background picture from the first version of my portfolio website. I like the design with the banner of info going through the middle of the background image, so I haven't changed much here. It's also mobile resopnsive.