Projects Section
At the moment I've hard-coded 5 projects to test the layout on different screen sizes. 3 Projects per screen width on desktop, 2 on tablet and 1 on mobile.

In my next commit I will fill out the array of objects (created in data.js) with the relevant info for each project I'll put on the website. After this I'll remove the hard-coded projects and map through the array of project objects to display each project on the page.

I've also put a white border-bottom on the navbar, so when the user scrolls through the page, the nabvar is seperated more clearly from each section.