Projects Display + Problem Solving Creation
Here I've finished filling out the data for the projects section. I've replaced the hard-coded projects with a map() function to display each project from the array of objects in data.js.

I've just created the necessary files for the ProblemSolving section. I'll continue to work on this section next.