Toggle Solution (ProblemSolving) & Footer component created
This commit has added the functionality to the ProblemSolving component. I gave the component some state for the problems and the toggleSolution function uses the SetProblems state.

The toggleSolution function maps through the prevState, and for each problem in the prevState, checks the index of the problem being clicked on for a match. When a match is found, the showingSolution property is toggled, when the index doesn't match it returns the problem. This is shown in the ternary operator.

Now when the button for each problem is clicked, the solution is show/hidden.

I've also created the footer component, I'll carry on with this in the next commit.