Project Links
Here I've added the links to the projects, so each live project and it's code is viewable.

I changed the ProjectBtn component from a <button> to an <a> in the ProjectElements page, and added the correct info for each mapped Project. (View Code Button)

For the link to each live project I wrapped the image of the project in an <a>, again giving each mapped Project all the necessary info.