Added new project / several small changes
Here I've removed the quiz from the Projects component and I've added the Mcfarlane's Coach Builders project.

I've added an offset to the smooth scroll for the nav links so the page scrolls to the correct section when a link is clicked. Before the links were scrolling slightly too far, hiding the title for each section.

I've also added the toggleMenu function to each NavLink, so now on mobile screens the menu closes when a link is clicked by the user.

I've also made the nav link container take up the full size of the screen when the menu is opened. I think it looks better than before.

I've changed the FooterEmail styled component from a h1 to an a tag. I've given it the relevant styling and it's now a clickable link to send an email to my emai address. Along with this I've changed the background color of the FooterContainerDisplay component to a dark blue.

I've also removed the default react favicon and replaced it with the screenshot of the website.